DOM manipulation Practice
Putting this lesson into practice.

Everything inside the "forEach" loop was removed (I left it commented out).

We create a <ul> element in the index.html as well as a <h1> element.

inside the "forEach" loop we created everything we have learned in this lesson. new variables mentioning the new elements of the index.html (for ul and li).
button-class comes from the css docuemnt. to give it more style and attractiveness.
all the "li", which are the pokemons that are inside the array were put in buttons with "listpokemon.appendChild(button)" and pokemonList which are the "li" were attached to "ul".

a new function was created inside the IIFE, "addListItem", everything created inside the forEach loop was cut and put inside this function.

Why?
When adding a new pokemon, a new list of the pokemon will be created automatically. Inside the loop where the information was cut from, now the new function will be running.