testing insert image with url
I have created two variables "newImg" which is inside addListItem. What I don't like is that the repeated image appears to me because I insert it inside "ul" with appendChild, depending on how many obejcts I have, so will be the number of repeated images.

the other variable is "newImage" and I insert it inside a div with #container, this form pleases me more because I could select the pokemon that I want to show.